[tests] re-orgranzie tests
