+++ hammerblade/torch/tests/targets.py
@@ -6,4 +6,5 @@
 """
 
 pytest_targets = [ "test_hammerblade_device",
-                   "test_hammerblade_kernels" ]
+                   "test_hammerblade_add_sub",
+                   "test_hammerblade_mul"]
